requirements.txt
aiohttp==3.9.3
aiosignal==1.4.0
alembic==1.13.1
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.21.3
anyio==4.12.1
asttokens==3.0.1
async-timeout==4.0.3
attrs==25.4.0
black==24.3.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
click==8.1.8
cryptography==46.0.5
decorator==5.2.1
distro==1.9.0
edge-tts==7.2.7
elevenlabs==1.0.0
exceptiongroup==1.3.1
executing==2.2.1
filelock==3.19.1
frozenlist==1.8.0
fsspec==2025.10.0
google-api-core==2.30.0
google-api-python-client==2.122.0
google-auth==2.49.1
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-cloud-texttospeech==2.16.3
googleapis-common-protos==1.73.0
greenlet==3.2.5
groq==0.4.2
grpcio==1.78.0
grpcio-status==1.62.3
h11==0.16.0
hf-xet==1.4.2
httpcore==1.0.9
httplib2==0.31.2
httpx==0.27.0
huggingface_hub==1.7.1
idna==3.11
iniconfig==2.1.0
ipython==8.18.1
jedi==0.19.2
loguru==0.7.2
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.3
matplotlib-inline==0.2.1
mdurl==0.1.2
multidict==6.7.1
mypy_extensions==1.1.0
oauthlib==3.3.1
openai==1.14.0
packaging==26.0
parso==0.8.6
pathspec==1.0.4
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.4.0
pluggy==1.6.0
prompt_toolkit==3.0.52
propcache==0.4.1
proto-plus==1.27.1
protobuf==4.25.8
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.6.3
pydantic-settings==2.2.1
pydantic_core==2.16.3
pydub==0.25.1
Pygments==2.19.2
pyparsing==3.3.2
pytest==8.1.1
pytest-asyncio==0.23.5
python-dateutil==2.9.0
python-dotenv==1.0.1
python-telegram-bot==21.0.1
PyYAML==6.0.3
redis==5.0.3
requests==2.31.0
requests-oauthlib==2.0.0
rich==14.3.3
rq==1.16.1
ruff==0.3.3
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.28
stack-data==0.6.3
tabulate==0.9.0
tenacity==8.2.3
tokenizers==0.22.2
tomli==2.4.0
tqdm==4.67.3
traitlets==5.14.3
typer==0.23.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
wcwidth==0.6.0
websockets==15.0.1
yarl==1.22.0
